Remove unused Arrays class and switch to faster array_map_assoc impl.
