+++ tests/benchmarks/class-functions-bench.php
@@ -228,6 +228,20 @@ class Functions_Bench {
 	}
 
 	/**
+	 * Provides an array_map implementation with control over resulting array's keys.
+	 *
+	 * @param  callable $callable A callback function that needs to $key, $value pairs.
+	 *                            The callback should return tuple where the first part
+	 *                            will be used as the key and the second as the value.
+	 * @param  array    $array    The array.
+	 *
+	 * @return array
+	 */
+	public static function array_map_assoc( callable $callable, array $array ) {
+		return array_column( array_map( $callable, array_keys( $array ), $array ), 1, 0 );
+	}
+
+	/**
 	 * Input data for array_manipulate tests.
 	 *
 	 * @return array
@@ -284,7 +298,7 @@ class Functions_Bench {
 	 * @param  array $params The parameters.
 	 */
 	public function bench_array_manipulate( $params ) {
-		\PHP_Typography\Arrays::array_map_assoc( function( $key, $value ) {
+		self::array_map_assoc( function( $key, $value ) {
 			return [ $value => $key ];
 		}, $params['array'] );
 	}
