+++ tests/fixes/node-fixes/class-smart-fractions-fix-test.php
@@ -36,7 +36,6 @@ use \PHP_Typography\U;
  *
  * @uses ::__construct
  * @uses PHP_Typography\Fixes\Node_Fixes\Abstract_Node_Fix::__construct
- * @uses PHP_Typography\Arrays
  * @uses PHP_Typography\DOM
  * @uses PHP_Typography\Settings
  * @uses PHP_Typography\Settings\Dash_Style
