+++ tests/fixes/node-fixes/class-space-collapse-fix-test.php
@@ -35,7 +35,6 @@ use \PHP_Typography\Settings;
  *
  * @uses ::__construct
  * @uses PHP_Typography\Fixes\Node_Fixes\Abstract_Node_Fix::__construct
- * @uses PHP_Typography\Arrays
  * @uses PHP_Typography\DOM
  * @uses PHP_Typography\Settings
  * @uses PHP_Typography\Settings\Dash_Style
