+++ tests/fixes/node-fixes/class-style-numbers-fix-test.php
@@ -38,7 +38,6 @@ use \PHP_Typography\Settings;
  * @uses PHP_Typography\Fixes\Node_Fixes\Abstract_Node_Fix::__construct
  * @uses PHP_Typography\Fixes\Node_Fixes\Classes_Dependent_Fix::__construct
  * @uses PHP_Typography\Fixes\Node_Fixes\Simple_Style_Fix::__construct
- * @uses PHP_Typography\Arrays
  * @uses PHP_Typography\DOM
  * @uses PHP_Typography\Settings
  * @uses PHP_Typography\Settings\Dash_Style
